JBPM-7405 fix tests after refactor on lienzo to change packages
