+++ src/test/java/com/ait/lienzo/client/core/shape/wires/MagnetManagerTest.java
@@ -22,8 +22,8 @@ import com.ait.lienzo.client.core.Context2D;
 import com.ait.lienzo.client.core.shape.Circle;
 import com.ait.lienzo.client.core.shape.IPrimitive;
 import com.ait.lienzo.client.core.shape.MultiPath;
-import com.ait.lienzo.client.core.shape.decorator.IShapeDecorator;
-import com.ait.lienzo.client.core.shape.decorator.MagnetDecorator;
+import com.ait.lienzo.client.core.shape.wires.decorator.IShapeDecorator;
+import com.ait.lienzo.client.core.shape.wires.decorator.MagnetDecorator;
 import com.ait.lienzo.client.core.util.ScratchPad;
 import com.ait.lienzo.test.LienzoMockitoTestRunner;
 import com.ait.tooling.nativetools.client.collection.NFastStringMap;
