+++ src/test/java/com/ait/lienzo/client/core/suite/WiresTestSuite.java
@@ -17,8 +17,6 @@
 package com.ait.lienzo.client.core.suite;
 
 import com.ait.lienzo.client.core.shape.AbstractOffsetMultiPointShapeTest;
-import com.ait.lienzo.client.core.shape.decorator.MagnetDecoratorTest;
-import com.ait.lienzo.client.core.shape.decorator.PointHandleDecoratorTest;
 import com.ait.lienzo.client.core.shape.wires.BackingColorMapUtilsTest;
 import com.ait.lienzo.client.core.shape.wires.MagnetManagerTest;
 import com.ait.lienzo.client.core.shape.wires.SelectionManagerTest;
@@ -27,6 +25,8 @@ import com.ait.lienzo.client.core.shape.wires.WiresContainerTest;
 import com.ait.lienzo.client.core.shape.wires.WiresManagerTest;
 import com.ait.lienzo.client.core.shape.wires.WiresShapeControlHandleListTest;
 import com.ait.lienzo.client.core.shape.wires.WiresShapeTest;
+import com.ait.lienzo.client.core.shape.wires.decorator.MagnetDecoratorTest;
+import com.ait.lienzo.client.core.shape.wires.decorator.PointHandleDecoratorTest;
 import com.ait.lienzo.client.core.shape.wires.handlers.WiresShapeHandlerImplTest;
 import com.ait.lienzo.client.core.shape.wires.handlers.impl.AlignAndDistributeControlImplTest;
 import com.ait.lienzo.client.core.shape.wires.handlers.impl.WiresCompositeControlImplTest;
